Renamed ResolverContainer#add(Object[, Closure]) to addLast.
This is for an upcoming change where the ResolverContainer will implement Collection, and this existing add() method is incompatible with the add() method from the Collection contract.

This is a breaking change, which is unavoidable. However, it's a level lower than what user's typically interact with so the impact is negligible.