Initial work on creating “NativeSourceSet” which is a NamedDomainObjectContainer of SourceDirectorySet implementations.
This is due to a native compile operation potentially being the combination of multiple different sets of different kinds of files to the compiler. What's to come is that each source directory set has to in some (abstract) way add itself to the compiler (which will end up as command line args).