Change the JVM's working dir much earlier in the command execution pipeline so that any relative paths sent by the client are resolved relative to the client's working dir.
This is a temporary fix. This means that we attempt all the JNA magic before we start routing logging back to the client, which isn't going to help debugging in the wild. The longer term fix will be to ship the desired log level to the daemon on it's own so logging can be established very early.