Support a build system property to place the daemon registry used for daemon integ tests in "$rootProject.buildDir/daemon".
This is so we can grab the daemon logs as build artifacts, and so they don't keep piling up on the CI server. We were storing thing in “intTestHome” in the project root which is never cleaned up. This is also made it difficult to grab the logs for that specific build.