Can use concurrent test util without extending ConcurrentSpecification
Made the ConcurrentSpecification a dummy class that simply contains a @Delagate to ConcurrentTestUtil instance. The latter one can be used directly in junit tests, or can be used as a junit @Rule, or can be used as @Delegate (or combinations of them). Later, we might get rid of ConcurrentSpecification completely and change all tests to use the rule & the delegate but for now, I want to make sure the rule+delegate pattern does the trick.