Change the daemon protocol to include an acknowledgement that the build has been accepted.
This should solve some intermittent failures we have been seeing with IO. What was happening previously was that input forwarding was being established with daemons before we knew whether they were busy or not. This ultimately meant we would suck up stdin and throw it at the daemon, which would just respond with “busy”. We'd then find a usable daemon, but the stdin had been consumed.

Now, we wait until the daemon acknowledges that it has “accepted” the build before consuming stdin.