GRADLE-2035 Fixed the "couldn't find test with id" problem...
Apparently, we can receive an output event for a test that we already received a complete event. However, I couldn't reproduce it in an integration test nor in local testing (I've generated hundreds of test classes, each with good number of test methods and ran it in various ways but the problem was not exposed). So, it's still a working theory. Also, we should peer review this implementation and try to find a much better solution.