Improving the tooling api coverage for concurrency.
The test that was exercising the concurrent scenario with different target gradle versions now truly uses the 'previous' version, instead of hardcoded M7. Introduced ReleasedVersions type that holds that information (this can be improved further). Unignored the concurrent test because the reason for failure we a red herring (I think).