Improved the diagnostics around daemon problems by fixing some logging problems...
Now the daemon infrastructure always logs to the deamon log file with DEBUG. This should be fine because the build requests carry the log level anyway, and for the execution of the build the requested log level is used. When the build finishes, though, the infrastructure is configured back to use DEBUG. Previous behavior was awkward because the infrastructure initially logged with DEBUG and then kept using the next build request's after the first build. Also made sure that the logging statements has proper format, e.g. contains the class name, etc. Added corresponding coverage. Might be worth doing the same for foreground daemon for consistency.