GRADLE-2099 - When starting a daemon, set -Dfile.encoding to the default encoding of the client jvm.
This is an interim/partial solution to make the daemon work with our build. The groovydoc task can only use the platform encoding which we need to be utf8, and we check for this. Because the daemon starter strips out any system properties that the daemon client was started with, the daemon is *always* going to use the platform's default encoding. This means that it's impossible to use the daemon with m8 and our build (if docs:groovydoc is going to be run).