GRADLE-2120,GRADLE-2119 - change configure by map algorithm to work with enhanced dynamic objects.
if the object has a property with the name it is set. If not, a one arg method with that name is tried. If that method doesn't exist, we set the property.

Previously, we tried to set the property first which meant that for enhanced objects the method version would never be tried. Instead, a new dynamic property would be created.