Add 1.8 value to the list of known Java versions
It is not currently possible to set source version to 1.8 with the Java plugin (you would get an index-out-of-bounds
exception).

With the initial OpenJDK 8 Project Lambda preview release [1], it would be nice to be able to experiment with
OpenJDK 8 using Gradle. As such, I decided to go ahead and add 1.8 to the list of known versions.

1. http://jdk8.java.net/lambda/