Fixed the issue that was blocking the wrapper upgrade.
This is tricky. BootstrapSecurityManager is hard to unit test because this is one of the classes that leaks to test vm. This is a short-term solution - this change needs to be applied along with wrapper upgrade so that BootstrapSecurityManagerTest passes from the terminal. Long term we need to figure out how to unit test this class.