Change the way we version the build to not use symbolic numbers by default.
By default, the version number will be “0.0-«timestamp»”, where «timestamp» is the last modified time of the most recently edited source file in the core project.

The idea is that now this build just builds Gradle. Allocating a symbolic version number to a build is done externally by the separate promotion process.