GRADLE-2244 - try to make cleaning more robust on Windows.
This is based on code in Ant that deals with deleting files. It eludes to there being a bug in Windows JDKs that is “fixable” by running a GC when a file files to delete and then trying again. Some googling finds similar information, but I haven't been able to find a definitive reference.

There is no added coverage for this as I'm unsure how to write a reliable test at this time. It can be manually tested though.