Dependency report housekeeping.
The new dependency report shows unresolved dependencies that the old report didn't show (see GRADLE-2410). I made the report look tiny bit better when there are some unresolved dependencies. I know we are not yet playing the 'shows unresolved dependencies' story. However, the report didn't not look well when there are dependencies with empty configurations (it is confusing and looks kind of like a bug). I wanted to fix that before the release.