Added ScalaIDE support to the Eclipse plugin.
The ScalaIDE classpath container is now added when to projects using
the Scala plugin. This resolve build problems after importing a Scala
project due to the ScalaIDE not being able to add its dynamically
computed entries to the classpath of the project.

The ScalaIDE adds the Scala libraries to the classpath, which
duplicates the dependencies provided by Gradle. In the ScalaIDE v2.0.2,
having multiple entries for scala-library results in full builds
instead of incremental). A detatched configuration allows projects to
exclude the dependencies from being provided by Gradle, which matches
the SBT and Maven behavior.