Introduce IvyNormalizedPublication and IvyPublisher into core, and enable the creation of Publishers from IvyArtifactRepository that can publish Ivy publications.
The goal here was to try and cause as little change to existing mechanisms as possible, but enable the new API. As such, the sharing of ArtifactPublisher impls (which backs IvyPublisher) introduced some nasty spaghetti. This will be detangled as ArtifactPublisher and its internals gets restructured to orient towards the new API.

Also, the contract of ArtifactRepositoryInternal has changed in that it now provides Publisher impls based on type tokens. This allows asserting that you have a compatible repository before normalizing the publication.