Don't use a RepositoryHandler for publishing.repositories, use a named object container of ArtifactRepository.
This avoids a bunch of awkwardness that stems from the name of the repository being far more important in publication than consumption. It also avoids propagating the RepositoryHandler API to a new place.

When just the “publishing” plugin is applied, repositories cannot be created. When the “ivy-publish” plugin is applied, the container creates instances of IvyArtifactRepository. This is ok for the initial release because that's the only type that we can deal with. Looking forward, we'll need to expand the domain container stuff to be polymorphic.