Fetch the necessary bits from publication services on demand.
This completely invalidated the approach used to unit test this task, so I've removed the test. It was of dubious value now that we have sufficient integration test coverage.

To effectively unit test this guy, we'd need to be able to construct with some mock services. This is not possible at the moment, but would be a great mechanism to have.