GRADLE-2408 - when using the daemon/tooling api, use the client's “GRADLE_HOME” as the location to find distribution init scripts.
A method getAllInitScripts() has been added to start parameter that provides the explicit (i.e. -I) and implicit (i.e. in gradle_home and gradle user home), and this is now used by the “server” instead of the server going looking for them (based on its GRADLE_HOME).

Nothing has been added to the release notes at this time. Not sure on whether this “bug fix” warrants special mention.