Add buildSetup customization support by adding '--type' Commandline option. - Add java-library as first type implementation - We manage supported types at the moment statically in ProjectLayoutSetupRegistryFactory
We currently evaluate the --type commandline option when populating the taskgraph.
This is to late to configure the plugin depending on the '--type' option.

In the current implementation of BuildSetupPlugin we evaluate the '--type' manually in the Plugin,
to allow configuration depending on this option.

Thought about creating ProjectLayoutSetupRegistryFactory in ProjectInternalServiceRegistry. Skipped that to keep all buildSetup classes in buildSetup subproject for now.