Ensure that we close what is backing the fake system input stream during a daemon build.
This was not being closed under all circumstances. Previously this was only being closed when the std in handler received a message from the client that there was no more input coming. This message could be sent _after_ the std in handler had been registered from handling such events on the daemon side.