REVIEW-3319: completely reworked TemplateBasedProjectInitDescriptor implementation
TemplateBasedProjectInitDescriptor now takes a set of TemplateOperation to generate a specific projectlayout. A SimpleTemplateOperation just contains the a source url, a target file and optinoal bindings
This makes composing different descriptors out of different templtae operations cleaner and simple. ProjectLayoutSetupRegistryFactory uses TemplateOperationBuilder to siimplify creating instances of TemplateOperation which are composed to different ProjectInitDescriptors