Added plugins.withId("pluginId") api for filtering plugins by id.
Needs review because I've used Closure type for method parameters. I've been following the pattern we already have for the PluginContainer. Can we start decorating the PluginContainer? Is it a breaking change? If so, let's do it in 2.0.

The api needs general review (e.g. return types, generic types, etc.)