FIx int test on Windows by not using a JRE as the alternative Java home.
This test was failing because the daemon that was launched was using an unexpected java home.
 The gradle.properties file was pointing to a JRE and the daemon infrastructure was expecting a daemon to start using this as the JAVA_HOME.
However, the daemon process determines its java home by way of Jvm.current().javaHome.
This, by way of findJavaHome() and findToolsJar(), actually reports the JDK for the same Java version as the Java home if it happens to exist in the same dir as the JRE.

Changing this test may not be the final answer.
This problem could also affect users if they happen to use org.gradle.javahome to point to a JRE when the corresponding JDK is available alongside it.
It's not clear to me why we are reporting the JDK as the Java home when the JRE has been used in the first place.